+++ src/aria2tui/aria2tui_app.py
@@ -2,6 +2,7 @@
 
 import os
 import sys
+from sys import exit
 import tempfile
 import time
 import toml
@@ -9,22 +10,24 @@ import json
 import curses
 
 os.chdir(os.path.dirname(os.path.realpath(__file__)))
-sys.path.append(os.path.expanduser("../../list_picker/"))
+os.chdir("../..")
+sys.path.append(os.path.expanduser("../../"))
 
-from list_picker.ui.list_picker_colours import get_colours, get_help_colours, get_notification_colours
-from list_picker.utils.options_selectors import default_option_input, output_file_option_selector, default_option_selector
-from list_picker.utils.table_to_list_of_lists import *
-from list_picker.utils.utils import *
-from list_picker.utils.sorting import *
-from list_picker.utils.filtering import *
-from list_picker.ui.input_field import *
-from list_picker.utils.clipboard_operations import *
-from list_picker.utils.searching import search
-from list_picker.ui.help_screen import help_lines
-from list_picker.ui.keys import list_picker_keys, notification_keys, options_keys, menu_keys
-from list_picker.utils.generate_data import generate_list_picker_data
-from list_picker.utils.dump import dump_state, load_state, dump_data
-from list_picker.list_picker_app import Picker, start_curses, close_curses
+# from list_picker.ui.list_picker_colours import get_colours, get_help_colours, get_notification_colours
+# from list_picker.utils.options_selectors import default_option_input, output_file_option_selector, default_option_selector
+# from list_picker.utils.table_to_list_of_lists import *
+# from list_picker.utils.utils import *
+# from list_picker.utils.sorting import *
+# from list_picker.utils.filtering import *
+# from list_picker.ui.input_field import *
+# from list_picker.utils.clipboard_operations import *
+# from list_picker.utils.searching import search
+# from list_picker.ui.help_screen import help_lines
+# from list_picker.ui.keys import list_picker_keys, notification_keys, options_keys, menu_keys
+# from list_picker.utils.generate_data import generate_list_picker_data
+# from list_picker.utils.dump import dump_state, load_state, dump_data
+# from list_picker.list_picker_app import Picker, start_curses, close_curses
+from list_picker.list_picker_app import *
 
 from aria2tui.lib.aria2c_wrapper import *
 from aria2tui.utils.aria2c_utils import *