+++ src/aria2tui/graphing/speed_graph.py
@@ -1,9 +1,13 @@
 import sys, os
-sys.path.append(os.path.expanduser("../../../list_picker/"))
-sys.path.append(os.path.expanduser("../../../aria2tui/"))
+# sys.path.append(os.path.expanduser("../../../list_picker/"))
+# sys.path.append(os.path.expanduser("../../../aria2tui/"))
+os.chdir(os.path.dirname(os.path.realpath(__file__)))
+os.chdir("../../..")
 from aria2tui.lib.aria2c_wrapper import *
 from aria2tui.utils.aria2c_utils import *
-from list_picker.list_picker_app import Picker, start_curses, close_curses
+# from list_picker.list_picker_app import Picker, start_curses, close_curses
+from list_picker import *
+from list_picker.list_picker_app import *
 import time
 import re
 from typing import Callable